Fix number of plural forms in files that don't have the plural forms header
lupdate creates .pot files without the Plural-Forms: line, when opening them lokalize ends up in this "if branch" where it adds the line to the header it will write when saving but it does not update the numberOfPluralForms variable so fields with plurals don't show with multiple options to translate (until you modify the file, save it, and open it again since then lokalize will have added the header and it'll work).

With this fields do properly shown on first open.

Acked by Nick, Luigi and Chusslove

REVIEW: 118006