appdata: use the screenshot from www.kde.org
This is the canonical place used for the official screenshots.
Moreover the old link points to a wiki page, not directly to the image,
so it had to be changed anyway or `appstream-cli validate-relax ...`
would not pass.

(cherry picked from commit feb56a609593ad4aa57b198ba9881f9f8eb9fd44)