cmake: rewrite FindHUNSPELL.cmake to use pkg-config
Rewrite the FindHUNSPELL.cmake, mainly to make use of pkg-config:
this way it is easier to adapt to new upstream versions, as the library
name is provided by the pkg-config file.  Also, the switch to
pkg-config gives a different include directory (i.e. directly where
hunspell.hxx is), so adapt lokalize accordingly.

Other changes to the module are:
- drop HUNSPELL_DEFINITIONS, unused
- HUNSPELL_INCLUDE_DIR -> HUNSPELL_INCLUDE_DIRS, like usual convention
- remove the silent/cache handling, it is not needed
- lowercase all the cmake commands
- pick the version from pkg-config