Right-align all translation count columns in Project Overview
Summary:
Before this patch, all columns in the Project Overview tab
were left-aligned. The patch changes the alignment for the
four translation count columns (‘Total’, ‘Translated’, ‘Not ready’
and ‘Untranslated’) to right alignment.

The effect is that the counts are easier to compare across files,
since the unit, ‘tens’, ‘hundreds’ etc. parts of the numbers are
now aligned.

BUG: 394512
FIXED-IN: 18.04.2

Test Plan:
  # Open a localisation project
  # Observe that four translation count columns (‘Total’, ‘Translated’, ‘Not ready’ and ‘Untranslated’) are now right-aligned, while they used to be left-aligned.
  # Observe that no other columns change alignment.

Reviewers: ltoscano, shaforostoff, sharvey, ngraham, aacid, #localization

Reviewed By: aacid, #localization

Subscribers: oysteins, kde-i18n-doc

Tags: #localization

Differential Revision: https://phabricator.kde.org/D13098