Show number of incomplete (fuzzy + untranslated) strings in project view
Summary:
This adds a new column, ‘Incomplete’ to the project view. This
shows the number of entries that are *either* fuzzy (‘not ready’)
or untranslated.

A use case might be someone updating a module/package which is *mostly*
translated, but which have seen some changes lately, with many files
with a *few* untranslated/fuzzy strings each (‘kde-workspace’ is a
typical example). The translators wants to quickly find which files
need updating. Sorting by the ‘Not ready’ column misses the files
with only *new* (‘Untranslated’ strings). Sorting by the ‘Untranslated’
column misses the files with only fuzzy (‘Not ready’) strings. So
before this patch, the translator needed to resort the list of files
twice (or actually four times, since clicking on a column first sorts
the list in ascending order, i.e. with the fully translated files first).
With the patch, the translator can easily sort the list so that it shows
*all* the files which need updating at the top, which makes their
work easier. :)

Test Plan:
- Open project view before and after patch.
- Notice new ‘Incomplete’ column.
- Sort the list of files by the ‘Incomplete’ column and check that it works.
- Check that the number in the ‘Incomplete’ column is the sum of the numbers
  in the ‘Not ready’ and ‘Untranslated’ columns.

Reviewers: #localization, aacid, ltoscano, shaforostoff

Reviewed By: #localization, aacid

Subscribers: #localization, kde-i18n-doc

Tags: #localization

Differential Revision: https://phabricator.kde.org/D13411