Fix drag'n'drop behavior within the Target area
Summary:
Fix the drag'n'drop behavior within the Target area

The behavior of Drag and drop within the Target area is currently very buggy.
The reason is that when you move some text from the middle of the sentence to a different place, the cursor loses the information of the selection (the cursor info becomes the end location of the selection and there is no storage of what the selection was).

To reproduce this move something in the middle of the sentence to the end of the sentence, you will see that actually your selection is just copied right after the selection.

In order to fix this behavior and make it move the text instead of just copying, this patch stores the original location of the selection, correctly updates the location of the cursor to the destination and deletes the text at the original location (except if you do Ctrl+Dragndrop).

https://bugs.kde.org/show_bug.cgi?id=346784

Reviewers: #localization, aacid, ltoscano, shaforostoff, mlaurent

Tags: #localization

Differential Revision: https://phabricator.kde.org/D14177