Allow lokalize to build on Windows with craft
Summary:
Allow lokalize to build on Windows with craft

With a few tweaks, it is possible to build lokalize on Windows with craft.
This generates a binary with KDE features (installer weight 100 Mb).

This patch allows building on Windows only with Craft with KF5, it doesn't fix the NOKDE (qt only) flag which seems very broken at the moment.

Reviewers: aacid, ltoscano, shaforostoff, mlaurent, #localization, huftis

Reviewed By: aacid, #localization

Tags: #localization

Differential Revision: https://phabricator.kde.org/D14321