Fix project model issues on file changes
Summary:
This revision fixes some issues with the Project Model refreshes
- Project view jumps to new (random?) file after closing translation file https://bugs.kde.org/show_bug.cgi?id=348757
- Whole project is refreshed when a file is changed https://bugs.kde.org/show_bug.cgi?id=343371
- Moving or adding a file/folder into the project root will trigger a bug where only that file/folder is displayed

It mostly fixes behaviors which are applied to directories while they shouldn't (like cache requests for directories that return empty values, or unnecessary full regeneration).

BUG: 348757
BUG: 343371

Test Plan:
I've tested the following test matrix :
* Move/Add/Remove/Save   x    File or Folder
* Within the root folder, within subfolders (same or different), between the root and a subfolder

This is for a .po project architecture, I would be interested to know if the behavior doesn't create regressions with a .po/.pot or other architecture (which might be more buggy to begin with).

Reviewers: ltoscano, #localization, aacid, mlaurent

Tags: #localization

Differential Revision: https://phabricator.kde.org/D15049