Improve management of empty/invalid files
Summary:
Previously lokalize SIGABRT when it encountered an invalid file within the project folder (for instance empty file).
Now the file will be shown with a red flag in the project view.
Empty files (no strings to translate) will also be shown with a yellow flag (the ability to filter out all empty files should make this manageable).

{F6219679}

Reviewers: ltoscano, aacid, mlaurent, #localization

Tags: #localization

Differential Revision: https://phabricator.kde.org/D15109