Simplify the behavior of tab switch navigation
Summary:
This feature simplifies the behavior of the tab switch navigation.

The Ctrl+[ and Ctrl+] shortcuts become Ctrl+Tab and Ctrl+Maj+Tab and override the previous default QMdiArea's embedded QTabBar shortcuts behavior (which only drew a box around the tab which would be selected once the key was released). The behavior will be similar to Chrome/Konsole by default (switch directly to the next/previous tab in the order the tabs are displayed). The exception will remain when a tab is closed, in that case the previously opened tab will be displayed.

This behavior can be overridden in the menu option by selecting "According to tab activation order". In that case the current behavior (prev tab selects the previously selected tab, next tab selects the oldest selected tab) will be used.

Finally a shortcut was added to just switch between the last two previously activated windows : Ctrl+[.

BUG: 357667

Reviewers: #localization, adrianchavesfernandez, huftis, aacid, ltoscano

Tags: #localization

Differential Revision: https://phabricator.kde.org/D15444