Add new ways to launch pology on file(s)
Summary:
Pology is a great tool but is not easily accessible from Lokalize currently, you need to launch a command line or configure a script, which is cumbersome and complicated in user space. After adding a line by line feature a few weeks ago, this review focuses on the file level :

- Launch the Pology "-s lokalize" command on file, files or folders from the Project Overview : this will open problematic messages in different tabs.
{F6285626}
- Launch the Pology "-s lokalize" command from within the file (Ctrl+Alt+P) : this will filter on problematic messages directly from within the file.
{F6285627}

The next step I see would be displaying pology statistics in the TM file, I'm open to your ideas for a better integration. I'm also counting on you to pass the message to your respective communities and update their documentation for the usage of pology.

As the agreement seems to be that syntax checks shall be handle externally, I will close the following bugs :
FEATURE: 342614
FEATURE: 342615
FEATURE: 65061
FEATURE: 198872
GUI:

Reviewers: #localization, aacid, adrianchavesfernandez, huftis, mlaurent

Reviewed By: mlaurent

Subscribers: mlaurent

Tags: #localization

Differential Revision: https://phabricator.kde.org/D15759