Bump lokalize version to 18.12
Summary:
I'm creating this revision to open a discussion on the lokalize version convention.

It is currently 2.0 both in the application and in the files generated (and has been like that since 2014), having it track the kde apps version would make bugs identification easier.

The only drawback is that the .po commits would be a bit more verbose if some people do not have the latest version.

Reviewers: #localization

Subscribers: ltoscano, yurchor

Tags: #localization

Differential Revision: https://phabricator.kde.org/D17364