Fix bug not finding local document
I set the source directory in lokalize to <src-dir> and tried to open a
file located in <src-dir>/xxx/yyy/zzz.cpp, but lokalize couldn't find
the file.

The problem is fixed with this patch by first searching the file under
the path formed by the source directory and the relative file path
from the po file.

CCBUG:403743
Reviewed By: #localization, sdepiets
Differential Revision: https://phabricator.kde.org/D19114