Display number columns in project overview using thousands separator
Before, the number of strings was displayed like, e.g. ‘12345’;
now it’s displayed like ‘12,345’, ‘12 345’ or ‘12.345’ (depending
on the user’s locale settings).

The ‘locale’ variable passed to PoItemDelegate::displayText()
by default was actually a *modified* version of the system locale,
one which omitted the thousands seperator, due to this Qt bug:

  https://bugreports.qt.io/browse/QTBUG-78094

As a temporary workaround, we therefore replace the ‘locale’
variable with the real system locale, i.e. QLocale::system().