Fix crash when a modified document and not saving changes
Summary:
If we don't delete the catalog explicitely it will be deleted in the middle of the deletion of EditorTab children.
That will cause the cleanChanged signal to be emmitted and then EditorTab::setModificationSign will be called.

Since we'll be in the middle of deleting EditorTab children very bad things can happen.

Reviewers: ltoscano, sdepiets, adrianchavesfernandez

Reviewed By: adrianchavesfernandez

Differential Revision: https://phabricator.kde.org/D24437