Add length columns to CatalogModelColumns
Summary:
This patch adds two columns to Translation Units toolview:
1) Source length,
2) Target length.
The columns display length of text in source and target columns
respectively.

The columns are useful in case of:
1) lots of translation units,
2) great variability in length between translation units.

The approach then is as follows:
1) sort translation units by source length column,
2) translate consecutively from the shortest to the longest source
string.

Advantage is in enabling translator to deliver more translated
translation units at given time compared to conventional approach:
1) sort translation units by entry column,
2) translate consecutively encountering mixed length source string.

The time to fully translate a file with translation units remains
unchanged.

Reviewers: ltoscano, sdepiets, aacid

Reviewed By: aacid

Differential Revision: https://phabricator.kde.org/D26477