Update mailings list for the Norwegian languages
Norwegian Bokmål and Norwegian Nynorsk has a new mailing list.
Northern Sami (‘se’) doesn’t really have a new mailing list,
but the old one is no longer available, and it’s acceptable to
use the mailing list for Norwegian Bokmål and Norwegian Nynorsk
for any queries about the Northern Sami translation (which for
the KDE software is currently not active).