Use thread specific names for connection Names
There seems to have been two issues with TM connections.
The first was that the TM connection was closed on project change (while they are actually all opened for stats purposes).
The second was that the connection names where not thread specific, which could cause issues, and it will now be the case.

BUG:422251
BUG:416427