Introduce SPDX conform copyright statements
This change extends the copyright extension logic to also detect SPDX-CopyrightText
statements for updating the copyright years upon saving. Secondly, the
serializing part is changed to store the copyright information as SPDX for
the current author.

Note that this change does not modify copyright statements form other authors,
which is important to allow a smooth transition for files where several
authors are editing the same file but with different versions of Lokalize.