Neaten layout of CatalogView / Translation Units
Uses spacing and margins to move the search box and button 6px away from the surrounding content, like in other parts of Lokalize.
Removes the border around the list of translation entries (this was causing a double border) and adds a horizontal line above the translation entries list to divide the search / button area from the list.
The option button title is capitalized to Option in line with other buttons in the UI.
Variables are renamed to clarify their function.

One issue is that the grey border still exists around the QDockWidget, you can see it to the left and right of the search / button area. When undocked, it also shows here. Below, where the list is, the border exists but is behind the list and the list extends to the edge of the widget i.e. over the border (and so with >0px margin around the list, the border is visible again). I have tried to figure out how to remove the outer border but haven't been able to. As a result, this merge request is an improvement but not necessarily a complete fix.

Before:

![Screenshot_20240723_055449](/uploads/0d6a4b43e5452c4b7adffbb876f6ef6e/Screenshot_20240723_055449.png)

After:

![Screenshot_20240723_072031](/uploads/87e33ebb8f69971b47c98522a210ddfa/Screenshot_20240723_072031.png)