Fix porting regression in diff computation
This relied on a non-standard matching behavior of 0x08 (backspace) in
QRegExp apparently, which PCRE/QRegularExpression doesn't have. So
manually replicate that by adding that to the pattern.

Add a unit test based on the original bug report, which now produces the
same result before and after porting.

Also needs backporting to 24.08.

cc @huftis

BUGS: 490837