Add action to skip HTML/XML tags
This can be used when editing strings containing
HTML, KUIT and other XML markup. The tags should
usually not be translated, so a shortcut for skipping
them is useful.

Newlines immediately after the end of a tag are also
skipped (by default, QRegularExpression doesn’t count
newlines as characters).

The shortcut is set to Ctrl+N by default, taking up
one of the (many) glossary insertion shortcuts.