Search across full relative paths of files and dirs in Project Overview
When using the search in Project Overview, if any part of the relative 
path from the project root to the file / dir matches, then the item is 
included in the search results: for a project with a tree containing 
'foo/bar.po', searching 'foo' will display 'foo/bar.po' and so will 
searching 'bar' or 'o/b' etc. This works into subdirectories, and 
also supports wildcards e.g. 'fo*.po' for our example above.

There are two other changes as well: enable alternating row colours 
because this is standard for other parts of Lokalize, and collapse 
directories when the search string is fewer than 3 characters, since 
we were already expanding directories by default for 3 or more chars.

BUG: 475163