Move TM and ATM entry metadata to third line, use QListWidget instead of QTextBrowser for entries
The same data is displayed per translation memory (TM) or alternate translation
memory (ATM) entry, but the data is ordered differently and formatted better.
An entry begins with metadata containing up to percentage match, number of
occurrences, and keyboard shortcut. The source and / or target strings follow
in a bullet point list. If the sources match by \>=95% then the N% match is in
bold to indicate a good match, on 100% match it says "perfect match" as well as
a percentage. If there is no keyboard shortcut associated with the entry then
no shortcut info is shown, rather than an empty placeholder. Proper Qt lists are
used for the entry items now rather than a text document. Where the ATM displays 
further details about the current translation rather than a translation from the 
alternate translation file, the ATM entry displays the details in the same 
format as the Unit Metadata window.