Use KMessageBox's persistent settings for the show ATV prompt
If alternate translation view data exists for an editor, then
prompt once per editor (per each restart of Lokalize) to show the
ATV. The prompt shows only once the related editor tab is clicked
on and only when an entry is shown that contains data for the ATV.
If the user ticks "Do not show again", then their choice is saved
permanently across Lokalize restarts and it is used to decide 
whether or not to automatically show the ATV when it contains data,
instead of prompting the user. This setting is saved in lokalizerc
as ResponseToShowAlternateTranslationsViewDialogWhenDataPresent
and replaces the related but different setting EverShown which made
the ATV show once, the very first time Lokalize was ever opened.