Sensible Project Overview column widths
This fixes the previous behaviour, to make it function as it 
was intended, with some improvements. The `lokalizerc` file 
is checked for the header state and restores column width and 
visibility of columns as it existed last time Lokalize was 
open. If the state is not found / fails to be applied, the 
default is used, which is the minimum width that will show 
the header title fully for most columns, and a semi-arbitrary 
wider width for columns with more text.

Previously the initial default column widths were too small, 
and the restore did not work (overrode column widths
regardless of the saved state, making them far too small).

Before:
![Screenshot__yyyy__MM__dd___HH__mm__ss_-1](/uploads/d5a70c4ec5043d8dce5aa5ab47007ff3/Screenshot__yyyy__MM__dd___HH__mm__ss_-1.png)

After:
![Screenshot__yyyy__MM__dd___HH__mm__ss_](/uploads/1aab75bf2dffcef0e7017db4d676367f/Screenshot__yyyy__MM__dd___HH__mm__ss_.png)