Fix editor tab state save / restore
There was a mix-up with converting to and from base64 encoding,
and more broadly there are two slightly conflicting state-saving
mechanisms, one using the field "DockWidgets" within a project
to store the editor tab's QMainWindow state, and the other using
the "EditorStates" global state storage section for project
agnostic editor tab states linked only to the file name the editor
tab is editing.

I've gotten rid of use of the "DockWidgets" fields in favour of the
"EditorStates" section in the hopes that this creates more robust
and longer-term state saving, and have fixed the state saving to
at least as good as it was a few months ago (issues still remain).

BUG: 501786

BUG: 412819
(cherry picked from commit 7286d371a0f969884ddd8800e1008dd11011dde5)