Convert b6a8be7f to use Settings::hideCompletedItems()
As discussed here https://invent.kde.org/sdk/lokalize/-/merge_requests/232#note_1211036

Adding HideCompletedItems to the lokalize.kcfg gives us
Settings::hideCompletedItems() getter, and adding it to
prefs_lokalize.kcfgc Mutators list gives us the setter
Settings::hideCompletedItems(). When the button is NOT
set to the default (i.e. when it is manually set to hide
fully translated files), the setting is saved in
~/.config/lokalizerc under the [General] category.