Improve HTML tag matching to reduce the false-positive colouring
By searching for tags with regex, we can reduce the number of false-
positive matches where text in the translation source / target is
coloured as though it were a HTML tag when it isn't. Previously, any
text with a `<` char in it would be coloured as HTML from that char,
now only from the `<` char followed by an alphabetic char,
potentially with a `/` char between i.e. matches `<s` in `<strong>`
and `</s` in `</strong>` but not something like `< b` or `<= 3`. In
my experience this is in line with how web engines parse HTML files.

Before:

![image](/uploads/5b7909579002204a6a94f3e798928831/image.png){width=675 height=590}

After:

![image](/uploads/ca5dfdde0decc98603231d41807439a7/image.png){width=681 height=557}

Original HTML tag highlighting is not changed:

![image](/uploads/910835acca0f003710547f87d6cc8585/image.png){width=232 height=92}

BUG: 504749