Open editor tab in background when Control key pressed
When the user left-clicks / presses Enter on a Project Overview item 
(item in ProjectWidget, which inherits QTreeView), open the file 
associated with the item in a new editor view. If the Control key is 
also pressed, do NOT focus the tab page (i.e. open editor tab in the
background, like internet browser behaviour) and instead remain on 
the Project Overview tab page.

I've also removed some qOverloads that were unnecessary because they
were applied to methods with no overloads.

BUG: 504957