doc: mention that Esc in catalog filter brings user back to TT field
This is so hard to notice that I only noticed it after reading the code
and saw that the filter's Escaped signal is hooked up to focus the Target
Text field. This came after I've prototyped an action to do the same,
thinking there is no keybind to jump from the filter back to the
Target Text field.

    // lokalize/src/editortab.cpp::221
    connect(m_transUnitsView, &CatalogView::escaped, this, &EditorTab::setProperFocus);

Nowhere in the interface is this indicated, even though it's very
useful. The analogous feature in browsers, Ctrl+l to focus the address
bar, does not leave the focus somewhere useful after Esc, so a user used
to Ctrl+l in browsers also wouldn't notice this feature.

This commit doesn't change that fact, but at least those who read the
manual get a chance to be taught this behavior right next to learning
that Ctrl+l focuses the filter bar.