Add KDDockWidgets as new dependency
Update: since the creation of this MR, KDAB released KDDW 2.3.0 
so the CI now uses that. This MR has been updated to reflect this.
For the CI to build Lokalize following this merge request, the 
following need merging:

KDDockWidgets update to 2.2.5 
https://invent.kde.org/packaging/craft-blueprints-kde/-/merge_requests/1246

Add KDDockWidgets to CI docker containers that Lokalize is built 
in https://invent.kde.org/sysadmin/ci-images/-/merge_requests/451