Add KDDockWidgets to kde-builder build dependencies
kde-ci.yml files are ingested by kde-builder tooling to create dependency
graphs, along with their CI functions. [1]
Commit 63766282 ("Add KDDockWidgets as new dependency") added a dependency on
KDDockWidgets in CMake, but did not add a corresponding dependency in
kde-ci.yml. This means that users would have to manually install KDDockWidgets
to build lokalize via kde-builder.
This is especially unfortunate on Ubuntu 25.10, which does not ship any version
of KDDockWidgets.
This commit adds the dependency to this project, so kde-builder will
build third-party/kddockwidgets before building lokalize when the
dependency file is regenerated.

This commit depends on the following change:
sysadmin/repo-metadata!660


[1] - https://community.kde.org/Infrastructure/Continuous_Integration_System